optimise cli's module dependencies (+ separate file for argparse)
