+++ abcd/frontends/cli/commands.py
@@ -1,79 +1,14 @@
-import re
 import logging
-
-from abcd import ABCD
-from pathlib import Path
-
-from abcd.frontends.cli.config import Config
-from abcd.frontends.cli.styles import SimpleStyle, FancyStyle
-
-from abcd.backends.abstract import URLError, AuthenticationError
-from abcd.parsers.arguments import key_val_str_to_dict
+from abcd.frontends.cli.decorators import init_style, init_db, init_config, check_readonly
 
 logger = logging.getLogger(__name__)
-
-
-def init_style(func):
-    def wrapper(*args, **kwargs):
-
-        style_str = kwargs.pop('style')
-        if style_str == 'simple':
-            style = SimpleStyle()
-        elif style_str == 'fancy':
-            style = FancyStyle()
-        else:
-            print('The style "{}" is not implemented!'.format(style_str))
-            exit(1)
-
-        func(*args, style=style, **kwargs)
-
-    return wrapper
-
-
-def init_config(func):
-    config = Config.load()
-
-    def wrapper(*args, **kwargs):
-        func(*args, config=config, **kwargs)
-
-    return wrapper
-
-
-def init_db(func):
-    def wrapper(*args, config, **kwargs):
-        url = config.get('url', None)
-
-        if url is None:
-            print('Please use abcd login first!')
-            exit(1)
-
-        db = ABCD(url=url)
-
-        default_query = kwargs.pop('default_query', [])
-        query = kwargs.pop('query', [])
-        q = default_query if default_query else []
-        q += query if query else []
-
-        func(*args, db=db, query=q, **kwargs)
-
-    return wrapper
-
-
-def check_readonly(func):
-    def wrapper(*args, **kwargs):
-        read_only = kwargs.pop('read_only')
-        if read_only:
-            print('In read only mode, you can\'t modify the data in the database')
-            exit(1)
-
-        func(*args, **kwargs)
-
-    return wrapper
 
 
 @init_config
 def login(*, config, name, url, **kwargs):
     logger.info('login args: \nconfig:{}, name:{}, url:{}, kwargs:{}'.format(config, name, url, kwargs))
+    from abcd import ABCD
+    from abcd.backends.abstract import URLError, AuthenticationError
 
     try:
         db = ABCD(url=url)
@@ -102,7 +37,6 @@ def download(*, db, query, filename, **kwargs):
     logger.info('download\n kwargs: {}'.format(kwargs))
 
     from ase.io import write
-
     write(filename, list(db.get_atoms(query=query)))
 
 
@@ -124,6 +58,8 @@ def delete(*, db, query, yes, **kwargs):
 @init_db
 @check_readonly
 def upload(*, db, path, extra_info, **kwargs):
+    from pathlib import Path
+
     path = Path(path)
     if path.is_file():
         db.upload(path, extra_info)
@@ -150,6 +86,8 @@ def summary(*, db, query, print_all, bins, truncate, style, props, **kwargs):
     if props is None:
         props_list = None
     else:
+        import re
+
         props_list = []
         for prop in props:
             # TODO: Check that is this the right place?
@@ -259,6 +197,8 @@ def summary(*, db, query, print_all, bins, truncate, style, props, **kwargs):
 @init_config
 @init_db
 def key_add(*, db, query, keys, **kwargs):
+    from abcd.parsers.arguments import key_val_str_to_dict
+
     data = key_val_str_to_dict(' '.join(keys))
 
     if db.count(query=query + [' or '.join(data.keys())]):
